flaskapp/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.2.0
pyzmq==19.0.1
Werkzeug==1.0.1